requirements.txt
Flask
twilio
python-dotenv
openai
requests
gunicorn
